Adição de tabelas e funcionalidades de login para aluno e professor

+++ src/main/java/grupo/quatro/api_manage_escola/Aluno/AlunoRepository.java
@@ -1,7 +1,14 @@
 package grupo.quatro.api_manage_escola.Aluno;
 
+import grupo.quatro.api_manage_escola.Usuario.UsuarioRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AlunoRepository extends JpaRepository<Aluno, Long> {
+import java.math.BigInteger;
+
+//public interface AlunoRepository extends JpaRepository<Aluno, Long> {
+//
+//}
+
+public interface AlunoRepository extends JpaRepository<Aluno, BigInteger> {
 
 }
